フィールド作成およびアクセサ作成ロジックをdtcパッケージに移動し、新しい抽象クラスAbstractDtcToolを追加
