+++ src/main/java/kmg/tool/application/logic/two2one/dtc/Enum2SwitchCaseCreationLogic.java
@@ -1,16 +1,16 @@
-package kmg.tool.application.logic.two2one;
+package kmg.tool.application.logic.two2one.dtc;
 
+import java.io.Closeable;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
 
-import kmg.tool.domain.logic.two2one.IctoOneLinePatternLogic;
 import kmg.tool.infrastructure.exception.KmgToolException;
 
 /**
- * <h2>メッセージの種類作成ロジックインタフェース</h2>
+ * <h2>列挙型からcase文作成ロジックインタフェース</h2>
  * <p>
- * メッセージの種類を定義するYMLファイルを自動生成するためのロジックインタフェースです。
+ * 列挙型の定義からswitch-case文を自動生成するためのロジックインタフェースです。
  * </p>
  *
  * @author KenichiroArai
@@ -19,7 +19,7 @@ import kmg.tool.infrastructure.exception.KmgToolException;
  *
  * @since 1.0.0
  */
-public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
+public interface Enum2SwitchCaseCreationLogic extends Closeable {
 
     /**
      * 項目名を書き込み対象に追加する。
@@ -49,7 +49,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      * @throws KmgToolException
      *                          KMGツール例外
      */
-    @Override
     boolean addOneLineOfDataToCsvRows() throws KmgToolException;
 
     /**
@@ -57,7 +56,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      *
      * @return true：成功、false：失敗
      */
-    @Override
     boolean clearCsvRows();
 
     /**
@@ -65,7 +63,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      *
      * @return true：成功、false：失敗
      */
-    @Override
     boolean clearProcessingData();
 
     /**
@@ -78,21 +75,20 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
     void close() throws IOException;
 
     /**
-     * メッセージの種類定義から項目と項目名に変換する。
+     * 列挙型定義から項目と項目名に変換する。
      *
      * @return true：変換あり、false：変換なし
      *
      * @throws KmgToolException
      *                          KMGツール例外
      */
-    boolean convertMessageTypesDefinition() throws KmgToolException;
+    boolean convertEnumDefinition() throws KmgToolException;
 
     /**
      * 変換後の1行データを返す。
      *
      * @return 変換後の1行データ
      */
-    @Override
     String getConvertedLine();
 
     /**
@@ -100,7 +96,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      *
      * @return 書き込み対象のCSVデータのリスト
      */
-    @Override
     List<List<String>> getCsvRows();
 
     /**
@@ -122,7 +117,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      *
      * @return 読み込んだ１行データ
      */
-    @Override
     String getLineOfDataRead();
 
     /**
@@ -138,7 +132,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      * @throws KmgToolException
      *                          KMGツール例外
      */
-    @Override
     boolean initialize(Path inputPath, Path outputPath) throws KmgToolException;
 
     /**
@@ -149,7 +142,6 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      * @throws KmgToolException
      *                          KMGツール例外
      */
-    @Override
     boolean readOneLineOfData() throws KmgToolException;
 
     /**
@@ -163,6 +155,5 @@ public interface MessageTypesCreationLogic extends IctoOneLinePatternLogic {
      * @throws KmgToolException
      *                          KMGツール例外
      */
-    @Override
     boolean writeCsvFile() throws KmgToolException;
 }
