+++ src/main/java/kmg/tool/presentation/ui/cli/two2one/dtc/AccessorCreationTool.java
@@ -1,23 +1,22 @@
-package kmg.tool.presentation.ui.cli.two2one;
+package kmg.tool.presentation.ui.cli.two2one.dtc;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
 
-import kmg.tool.application.service.two2one.Enum2SwitchCaseCreationService;
-import kmg.tool.presentation.ui.cli.AbstractDynamicTemplateConversionTool;
+import kmg.tool.application.service.two2one.AccessorCreationService;
 
 /**
- * <h2>列挙型からcase文作成ツール</h2>
+ * <h2>アクセサ作成ツール</h2>
  * <p>
- * 列挙型の定義からswitch-case文を自動生成するためのツールです。
+ * Javaクラスのフィールドに対するアクセサメソッド（getterおよびsetter）を自動生成するためのツールです。
  * </p>
  * <p>
- * このツールは入力ファイルとテンプレートファイルを使用して、switch-case文を含む出力ファイルを生成します。
+ * このツールは入力ファイルとテンプレートファイルを使用して、アクセサメソッドを含む出力ファイルを生成します。
  * </p>
  * <p>
- * AbstractDynamicTemplateConversionToolを継承しており、動的テンプレート変換処理を実装しています。
+ * AbstractTwo2OneToolを継承しており、2つの入力ファイルから1つの出力ファイルを生成する処理を実装しています。
  * </p>
  *
  * @author KenichiroArai
@@ -29,7 +28,7 @@ import kmg.tool.presentation.ui.cli.AbstractDynamicTemplateConversionTool;
 @SpringBootApplication(scanBasePackages = {
     "kmg"
 })
-public class Enum2SwitchCaseCreationTool extends AbstractDynamicTemplateConversionTool {
+public class AccessorCreationTool extends AbstractDtcTool {
 
     /**
      * <h3>ツール名</h3>
@@ -37,16 +36,16 @@ public class Enum2SwitchCaseCreationTool extends AbstractDynamicTemplateConversi
      * このツールの表示名を定義します。
      * </p>
      */
-    private static final String TOOL_NAME = "列挙型からcase文作成ツール";
+    private static final String TOOL_NAME = "アクセサ作成ツール";
 
     /**
-     * <h3>列挙型からcase文作成サービス</h3>
+     * <h3>アクセサ作成サービス</h3>
      * <p>
-     * 列挙型定義からswitch-case文を生成するためのサービスです。
+     * フィールド定義からアクセサメソッドを生成するためのサービスです。
      * </p>
      */
     @Autowired
-    private Enum2SwitchCaseCreationService enum2SwitchCaseMakingService;
+    private AccessorCreationService accessorCreationService;
 
     /**
      * <h3>エントリポイント</h3>
@@ -77,9 +76,9 @@ public class Enum2SwitchCaseCreationTool extends AbstractDynamicTemplateConversi
     public static void main(final String[] args) {
 
         @SuppressWarnings("resource")
-        final ConfigurableApplicationContext ctx = SpringApplication.run(Enum2SwitchCaseCreationTool.class, args);
+        final ConfigurableApplicationContext ctx = SpringApplication.run(AccessorCreationTool.class, args);
 
-        final Enum2SwitchCaseCreationTool tool = ctx.getBean(Enum2SwitchCaseCreationTool.class);
+        final AccessorCreationTool tool = ctx.getBean(AccessorCreationTool.class);
 
         /* 初期化 */
         tool.initialize();
@@ -94,33 +93,33 @@ public class Enum2SwitchCaseCreationTool extends AbstractDynamicTemplateConversi
     /**
      * <h3>コンストラクタ</h3>
      * <p>
-     * 列挙型からcase文作成ツールのインスタンスを生成します。
+     * アクセサ作成ツールのインスタンスを生成します。
      * </p>
      * <p>
      * 親クラスのコンストラクタを呼び出し、ツール名を設定します。 このコンストラクタによって、デフォルトのテンプレートパスも設定されます。
      * </p>
      */
-    public Enum2SwitchCaseCreationTool() {
+    public AccessorCreationTool() {
 
-        super(Enum2SwitchCaseCreationTool.TOOL_NAME);
+        super(AccessorCreationTool.TOOL_NAME);
 
     }
 
     /**
-     * <h3>列挙型からcase文作成サービスを返す</h3>
+     * <h3>アクセサ作成サービスを返す</h3>
      * <p>
-     * AbstractDynamicTemplateConversionToolの抽象メソッドを実装し、DI（依存性注入）された 列挙型からcase文作成サービスのインスタンスを返します。
+     * AbstractTwo2OneToolの抽象メソッドを実装し、DI（依存性注入）された アクセサ作成サービスのインスタンスを返します。
      * </p>
      * <p>
-     * このメソッドは親クラスの処理から呼び出され、実際のcase文生成処理を担当する サービスを提供します。
+     * このメソッドは親クラスの処理から呼び出され、実際のアクセサ生成処理を担当する サービスを提供します。
      * </p>
      *
-     * @return 列挙型からcase文作成サービス このツールが使用する列挙型からcase文作成サービスのインスタンス
+     * @return アクセサ作成サービス このツールが使用するアクセサ作成サービスのインスタンス
      */
     @Override
-    protected Enum2SwitchCaseCreationService getIoService() {
+    protected AccessorCreationService getIoService() {
 
-        final Enum2SwitchCaseCreationService result = this.enum2SwitchCaseMakingService;
+        final AccessorCreationService result = this.accessorCreationService;
 
         return result;
